backend/requirements.txt
# Backend dependencies
fastapi
uvicorn
qdrant-client
sentence-transformers
requests
python-dotenv

